API updates

+++ src/ServerFlex.API/Entities/PricingEntity.cs
@@ -9,13 +9,13 @@ namespace ServerFlex.API.Entities
         /// Gets or sets the hourly cost.
         /// </summary>
         [JsonProperty("costHourly")]
-        public double CostHourly { get; set; }
+        public decimal CostHourly { get; set; }
 
         /// <summary>
         /// Gets or sets the monthly cost.
         /// </summary>
         [JsonProperty("costMonthly")]
-        public double CostMonthly { get; set; }
+        public decimal CostMonthly { get; set; }
 
         /// <summary>
         /// Gets or sets the currency code, in ISO 4217 format.
